Feat: Point 3 ended

+++ src/main/java/com/travelport/models/SaleDetailDTO.java
@@ -3,7 +3,7 @@ package com.travelport.models;
 public class SaleDetailDTO {
 
     private Integer productCode;
-    private Integer quantity;
+    private Double quantity;
 
     public Integer getProductCode() {
         return productCode;
@@ -13,11 +13,11 @@ public class SaleDetailDTO {
         this.productCode = productCode;
     }
 
-    public int getQuantity() {
+    public Double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(Double quantity) {
         this.quantity = quantity;
     }
 }
